# src/model/dto/EmprestimoDto.ts
export class EmprestimoDto {
    id: number;
    livroId: number;
    usuarioId: number;
    dataEmprestimo: Date;
    dateDevolucao: Date;

    constructor(id: any, livroId: any, usuarioId: any, dataEmprestimo: any, dataDevolucao: any) {
        this.id = id;
        this.livroId = livroId;
        this.usuarioId = usuarioId;
        this.dataEmprestimo = dataEmprestimo;
        this.dateDevolucao = dataDevolucao;
    }
}   